add instance info cmd
